finance-platform/pyproject.toml
[project]
name = "finance-platform"
version = "0.1.0"
description = "Enterprise-style personal finance and expense intelligence API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "email-validator>=2.2.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<5",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.7",
  "celery>=5.4.0",
  "kafka-python>=2.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "httpx>=0.27.0",
  "ruff>=0.5.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
